Forgot to uncomment a line to compose nrr and aff in reg_transform

+++ nifty_reg/reg-lib/_reg_f3d2.cpp
@@ -99,7 +99,7 @@ int reg_f3d2<T>::AllocateCurrentInputImage(int level)
         this->controlPointGrid->pixdim[5]=this->stepNumber;
         this->controlPointGrid->du=this->stepNumber;
     }
-    else this->stepNumber=this->controlPointGrid->du;
+    else this->stepNumber=(int)this->controlPointGrid->du;
 
 #ifdef NDEBUG
     if(this->verbose){
